Fix Loop Hook
