+++ src/app/components/utility/utility.component.html
@@ -25,7 +25,11 @@
       @if (viewMode === ViewMode.Statistics) {
         <div>
           <div class="form">
-            <select class="form-select" id="ConstraintPropertySelect" [(ngModel)]="selectedStatistics">
+            <select
+              class="form-select"
+              id="ConstraintPropertySelect"
+              [(ngModel)]="selectedStatisticsId"
+              (ngModelChange)="changeSelectedStatisticsId()">
               @for (selectDataGroup of statisticsSelectData | groupBy: 'group'; track selectDataGroup) {
                 <optgroup [label]="selectDataGroup.key">
                   @for (selectData of selectDataGroup.value; track selectData) {
@@ -62,12 +66,12 @@
   <div
     class="utility-container d-flex overflow-y-scroll mt-2"
     [ngClass]="utilityContainerVisible ? 'visible' : 'hidden'">
-    @if (viewMode === ViewMode.Statistics) {
+    @if (viewMode === ViewMode.Statistics && selectedStatisticsFormatter) {
       <app-statistics
         class="d-flex flex-grow-1"
         [allocationData]="allocationData"
-        [selectedDataId]="selectedStatistics"
-        (selectDataChange)="statisticsSelectDataChange($event)"></app-statistics>
+        [selectedDataId]="selectedStatisticsId"
+        [formatter]="selectedStatisticsFormatter"></app-statistics>
     }
     @if (viewMode === ViewMode.Students) {
       @for (students of [allocationData.studentsWithoutTeam]; track students) {
