+++ src/app/components/utility/utility.component.ts
@@ -1,35 +1,50 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
 import { SelectData } from 'src/app/shared/matching/constraints/constraint-functions/constraint-function';
 import { AllocationData } from 'src/app/shared/models/allocation-data';
 import { facCloseIcon } from 'src/assets/icons/icons';
+import { IntroCourseProficiencyChartDataService } from '../statistics/charts/chart-data-formatter/formatters/intro-course-proficiency-chart-data.service';
+import { PriorityChartDataService } from '../statistics/charts/chart-data-formatter/formatters/priority-chart-data.service';
+import { SkillsProficiencyChartDataService } from '../statistics/charts/chart-data-formatter/formatters/skills-proficiency-chart-data.service';
+import { ChartDataFormatter } from '../statistics/charts/chart-data-formatter/chart-data-formatter';
 
 enum ViewMode {
   Students,
   Statistics,
 }
-
-export enum StatisticsViewMode {
-  PriorityDistribution = 'Priority Distribution',
-  Proficiency = 'Intro Course Proficiency',
-}
-
 @Component({
   selector: 'app-utility',
   templateUrl: './utility.component.html',
   styleUrl: './utility.component.scss',
 })
-export class UtilityComponent {
+export class UtilityComponent implements OnInit, OnChanges {
   facCloseIcon = facCloseIcon;
 
   viewMode = ViewMode.Statistics;
   ViewMode = ViewMode;
 
   statisticsSelectData: SelectData[] = [];
-  selectedStatistics: string = 'statistics-priority-distribution';
+  selectedStatisticsId: string;
+  selectedStatisticsFormatter: ChartDataFormatter;
 
   utilityContainerVisible = true;
 
   @Input({ required: true }) allocationData: AllocationData;
+
+  constructor(
+    private introCourseProficencyChartData: IntroCourseProficiencyChartDataService,
+    private priorityChartDataService: PriorityChartDataService,
+    private skillsProficiencyChartData: SkillsProficiencyChartDataService
+  ) {}
+
+  ngOnInit(): void {
+    this.updateSelectData();
+    this.changeSelectedStatisticsId();
+  }
+
+  ngOnChanges(): void {
+    this.updateSelectData();
+    this.changeSelectedStatisticsId();
+  }
 
   selectViewMode(viewMode: ViewMode): void {
     this.viewMode = viewMode;
@@ -38,11 +53,22 @@ export class UtilityComponent {
   toggleUtilityContainer(): void {
     this.utilityContainerVisible = !this.utilityContainerVisible;
   }
-
-  statisticsSelectDataChange(selectData: SelectData[]): void {
-    this.statisticsSelectData = selectData;
-    if (!!selectData && selectData.length > 0 && !this.selectedStatistics) {
-      this.selectedStatistics = selectData[0].name;
+  changeSelectedStatisticsId() {
+    if (!this.selectedStatisticsId) {
+      this.selectedStatisticsId = this.statisticsSelectData[0]?.id;
     }
+
+    this.selectedStatisticsFormatter = this.statisticsSelectData.find(
+      selectData => selectData.id === this.selectedStatisticsId
+    )?.reference;
+  }
+
+  updateSelectData(): void {
+    const formatters = [
+      this.priorityChartDataService,
+      this.introCourseProficencyChartData,
+      this.skillsProficiencyChartData,
+    ];
+    this.statisticsSelectData = formatters.flatMap(formatter => formatter.getSelectData());
   }
 }
